Refactor dashboard rendering to use view model presenter
